# Story desktop GPU memory optimization so videos play smoothly.