--- a/extensions/amp-story/1.0/amp-story-desktop-panels.css
+++ b/extensions/amp-story/1.0/amp-story-desktop-panels.css
@@ -89,7 +89,6 @@ amp-story[standalone].i-amphtml-story-desktop-panels {
   transform: scale(0.9) translateX(calc(250% + 192px)) translateY(0%) !important;
   opacity: 0.05 !important;
   transform-origin: left !important;
-  border-radius: 0 !important;
 }
 
 [dir=rtl] .i-amphtml-story-desktop-panels amp-story-page.i-amphtml-element {
@@ -185,8 +184,4 @@ amp-story[standalone].i-amphtml-story-desktop-panels {
   max-width: calc(3/5 * 75vh) !important;
   min-width: 320px !important;
   min-height: 533px !important;
-}
-
-.i-amphtml-story-desktop-panels > amp-story-page {
-  box-shadow: 0 50px 200px rgba(0, 0, 0, 0.7) !important
 }