The GPU memory used by our desktop three panels experience was so high that on some computers, the hardware-accelerated video decoding was stuttering, resulting on videos not playing.
(Hardware accelerated video decoding uses GPU, not CPU, which is why videos were playing smoothly when we were disabling it).

This fix gets all videos to play as smooth as expected in this desktop three panels layout, but as a nice side effect gets the overall experience much much faster.
Removing this `box-shadow` has no visible effect, since we removed the background images.

One line fix but really interesting performance improvements:

| Device  | Max GPU memory / Min FPS before | Max GPU memory / Min FPS after | Improvement |
| :-------------: | :-------------: | :-------------: | :-------------: |
| Dell workstation (linux)  | ~73mb / 41 FPS | 13mb / 56 FPS | 5.6x less GPU memory used |
| MacBook pro 2017 13i  | ~280mb / 17 FPS | ~60mb / 42 FPS | 4.7x less GPU memory used |

Fixes #23184