@gmajoulet 2019-07-09
`border-radius` is just some cleanup, it all came from the `box-shadow`