@newmuis 2019-07-10
FYI @hongwei1990 since we added (or, drastically increased) this when we removed the background images